# docs/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Procurement and PLS Seedbed API
  version: 0.24.0
  summary: Compliance-first procurement evidence API with proof anchoring and PLS seedbed support.
  description: >
    API contract for the supervisor-demo plus selected pilot-hardening baseline.
    The runtime uses opaque bearer sessions. OAuth2/OIDC entries are readiness
    boundaries and do not claim a production identity provider.
...
servers:
  - url: http://localhost:3100/api/v1
    description: Local backend
...
paths:
...
  /invoices:
    get:
      tags: [Invoices]
      operationId: listProcurementInvoices
      summary: List invoices visible to the authenticated actor.
      security:
        - bearerSession: []
      responses:
        "200":
          description: Visible invoices.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
    post:
      tags: [Invoices]
      operationId: submitProcurementInvoice
      summary: Submit invoice metadata for three-way matching.
      security:
        - bearerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId, amount, currency]
              properties:
                orderId: { type: string }
                deliveryEvidenceId: { type: string }
                amount: { type: string, example: "68000.00" }
                tax: { type: string }
                currency: { type: string, example: MYR }
                invoiceReference: { type: string }
                invoiceHash: { type: string, example: sha256:abc123 }
      responses:
        "200":
          description: Invoice submitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
  /invoices/{id}:
    get:
      tags: [Invoices]
      operationId: getProcurementInvoice
      summary: Get invoice metadata and match state.
      security:
        - bearerSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invoice metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
  /invoices/{id}/verify-match:
    post:
      tags: [Invoices]
      operationId: verifyProcurementInvoiceMatch
      summary: Verify invoice against order and delivery evidence metadata.
      security:
        - bearerSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invoice match verification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
  /invoices/{id}/approve-payment:
    post:
      tags: [Invoices]
      operationId: approveProcurementInvoicePaymentReadiness
      summary: Approve invoice for sandbox payment readiness without executing a payment.
      security:
        - bearerSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Invoice marked payment-ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
  /procurement-cases/{caseId}/closeout:
    post:
      tags: [Procurement Closeout]
      operationId: closeProcurementCase
      summary: Close a procurement case after evidence and invoice readiness review.
      security:
        - bearerSession: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes: { type: string }
      responses:
        "200":
          description: Procurement case closed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementCaseCloseoutResponse"
  /procurement-cases/{caseId}/summary:
    get:
      tags: [Procurement Closeout]
      operationId: getProcurementCaseSummary
      summary: Summarize case evidence, invoice, proof, and closeout state.
      security:
        - bearerSession: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Procurement case summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementCaseSummaryResponse"
  /suppliers/{supplierOrganizationId}/performance:
    get:
      tags: [Procurement Closeout]
      operationId: getSupplierPerformance
      summary: Get supplier scorecard metrics from procurement records.
      security:
        - bearerSession: []
      parameters:
        - name: supplierOrganizationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Supplier performance summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplierPerformanceResponse"
...
components:
  securitySchemes:
    bearerSession:
      type: http
      scheme: bearer
      bearerFormat: opaque-session
      description: Current runtime opaque bearer session token.
    oauth2:
      type: oauth2
      description: OAuth2/OIDC readiness boundary only; no production IdP is configured.
      flows:
        authorizationCode:
          authorizationUrl: https://identity.example.test/oauth/authorize
          tokenUrl: https://identity.example.test/oauth/token
          scopes:
            org:read: Read organization profile and context.
            org:write: Update organization metadata.
            org:users:manage: Manage organization users.
            network:read: Read organization network graph.
            network:write: Create and decide network requests.
            deals:read: Read company ledger projections.
            deals:write: Prepare deal actions.
            productivity:read: Read productivity workspace.
            productivity:write: Update tasks and saved views.
            proof:read: Read proof metadata.
            proof:verify: Verify proof hashes.
            email:read: Read local notification outbox.
            admin:platform: Platform administration.
...
  schemas:
...
    ProcurementInvoice:
      type: object
      properties:
        invoiceId: { type: string }
        orderId: { type: string }
        deliveryEvidenceId: { type: string }
        supplierOrganizationId: { type: string }
        buyerOrganizationId: { type: string }
        amount: { type: string, example: "68000.00" }
        currency: { type: string, example: MYR }
        invoiceReference: { type: string }
        invoiceHash: { type: string, example: sha256:abc123 }
        status: { type: string, enum: [submitted, matchPassed, matchFailed, paymentApproved, rejected] }
        matchResult: { type: object }
        lifecycleEventIds:
          type: array
          items: { type: string }
        latestLifecyclePayloadHash: { type: string }
    InvoiceResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ProcurementInvoice"
    InvoiceListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ProcurementInvoice"
    SupplierPerformanceSummary:
      type: object
      properties:
        supplierOrganizationId: { type: string }
        orderCount: { type: integer }
        deliveryEvidenceCount: { type: integer }
        invoiceCount: { type: integer }
        invoiceExceptionCount: { type: integer }
        proofCoveragePercent: { type: integer }
        closeoutCount: { type: integer }
        score: { type: integer }
        lastUpdatedAt: { type: string, format: date-time }
    ProcurementCaseCloseoutResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            closeoutId: { type: string }
            caseId: { type: string }
            orderId: { type: string }
            status: { type: string, enum: [open, closed] }
            metrics:
              $ref: "#/components/schemas/SupplierPerformanceSummary"
    ProcurementCaseSummaryResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          properties:
            caseId: { type: string }
            orderId: { type: string }
            sourceToAwardCaseId: { type: string }
            deliveryEvidenceCount: { type: integer }
            invoiceCount: { type: integer }
            invoiceExceptionCount: { type: integer }
            proofCoveragePercent: { type: integer }
            closeout: { type: object }
    SupplierPerformanceResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/SupplierPerformanceSummary"
...